+++ src/components/index.js
@@ -41,6 +41,7 @@ import LocationComponent from './location/Location';
 import FileComponent from './file/File';
 import TabsComponent from './tabs/Tabs';
 import TreeComponent from './tree/Tree';
+import ReCaptchaComponent from './recaptcha/ReCaptcha';
 import ModalEdit from './modaledit/ModalEdit';
 
 export default {
@@ -87,5 +88,6 @@ export default {
   location: LocationComponent,
   file: FileComponent,
   tree: TreeComponent,
+  recaptcha: ReCaptchaComponent,
   modaledit: ModalEdit,
 };